Refactor gasprice package
